Update inventory import front-end
