+++ app/javascript/packs/vue/import_repository_modal.js
@@ -1,7 +1,7 @@
 import PerfectScrollbar from 'vue3-perfect-scrollbar';
 import { createApp } from 'vue/dist/vue.esm-bundler.js';
 import 'vue3-perfect-scrollbar/dist/vue3-perfect-scrollbar.css';
-import ImportRepositoryModal from '../../vue/repositories/modals/import.vue';
+import ImportRepositoryModal from '../../vue/repositories/modals/import/container.vue';
 import { mountWithTurbolinks } from './helpers/turbolinks.js';
 
 const app = createApp({});
